% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extended_geneplot.R
\name{plot_geneplot_spdfs}
\alias{plot_geneplot_spdfs}
\title{Lower-level function for carrying out the plotting step of
\link{extended_geneplot}.}
\usage{
plot_geneplot_spdfs(
  logprob,
  spdf_vals1,
  spdf_vals2,
  plot_params,
  show_overlap_areas = TRUE,
  show_legend = TRUE,
  show_legend_below = TRUE,
  show_title = TRUE,
  title_text = NULL,
  show_quantiles = TRUE,
  only_plot_baseline_pop = FALSE,
  legend_width = 4,
  show_include_ids = FALSE
)
}
\arguments{
\item{logprob}{Log-Genotype Probabilities for individual genotypes with
respect to each of the reference populations. Required for GenePlot,
available as an output from \link{calc_geneplot_spdfs}.}

\item{spdf_vals1}{Genetic Distributions with first reference population as
the baseline population. Required for the auxiliary side plots, available as
an output from \link{calc_geneplot_spdfs}.}

\item{spdf_vals2}{Genetic Distributions with second reference population as
the baseline population. Required for the auxiliary side plots, available as
an output from \link{calc_geneplot_spdfs}.}

\item{plot_params}{Plotting options and settings, output from
\link{prepare_plot_params}. List includes \code{allpopnames, logten, split_dat,
colvec, shapevec, rimvec, orderpop, axispop, line_widths, cexpts,
mark_impute, xyrange, ylim_input, display_names, x_lgp_label, y_lgp_label,
grayscale_quantiles}.}

\item{show_overlap_areas}{(default FALSE) If TRUE, print the numerical
Overlap Area values beneath the extended plot.}

\item{show_legend}{(default TRUE) If TRUE, plot the legend.}

\item{show_legend_below}{(default TRUE) If TRUE, show the legend as a row
of coloured shapes and labels below the extended plot. If FALSE, show the
legend within the GenePlot. Default is TRUE because putting the legend
inside the central GenePlot can hide points from individual genotypes.}

\item{show_title}{(default TRUE) Include a title for the plot.}

\item{title_text}{(Optional) Provide alternative title text for the plot.}

\item{show_quantiles}{(default TRUE) If TRUE, plot quantiles of each baseline
population in the side plots.}

\item{only_plot_baseline_pop}{(default FALSE) Plot the baseline population in
each side plot i.e. the corresponding reference population, and do not plot
the comparison population.}

\item{legend_width}{(default 4cm) Change the width of the displayed legend.
Units are cm.}

\item{show_include_ids}{(default FALSE) Ignored unless \code{includepopnames}
is supplied. If TRUE, plots individual genotypes from the additional
included pops as points but also displays the ID for each genotype. This
can be useful if one or more genotype shows unusual patterns of fit to the
reference populations and you want to identify those genotypes.}
}
\value{
A list with the following components:

    \code{allpopnames} Vector of populations included in the plot, starting
    with the reference populations followed by the included populations.

    \code{logten} Whether the log scales in the plot are base 10 (default)
    or base e.

    \code{split_dat} The Log-Genotype Probabilities for the supplied genotypes,
    split up into a list of data frames, with each data frame corresponding
    to the genotypes from a single named population and containing their
    Log-Genotype Probabilities with respect to both of the reference
    populations.

    \code{colvec, shapevec, rimvec} Details of the colours and shapes used
    for the populations, and the outline colours for the GenePlot points

    \code{orderpop, axispop} Order in which the populations were plotted in
    the central GenePlot, from back to front, and the x-y order of the two
    reference pops.

    \code{line_widths} Line widths used in the side plot saddlepoint
    distribution curves.

    \code{cexpts} Point sizes used in the central GenePlot.

    \code{mark_impute, xyrange, ylim_input, display_names} Records the input
    options used to produce the plot.

    \code{x_lgp_label, y_lgp_label} Full axis labels used in the plot.

    \code{grayscale_quantiles} Whether the quantiles are plotted in grey or
    using the assigned population colours.
}
\description{
Use after running \link{calc_geneplot_spdfs} to
produce \code{logprob}, \code{spdf_vals1} and \code{spdf_vals2} and
\link{prepare_plot_params} to set up the plotting details.
}
